+++ pyslvs_ui/io/overview_ui.py
@@ -2,7 +2,7 @@
 
 # Form implementation generated from reading ui file 'pyslvs_ui/io/overview.ui'
 #
-# Created by: PyQt5 UI code generator 5.13.0
+# Created by: PyQt5 UI code generator 5.13.1
 #
 # WARNING! All changes made in this file will be lost!
 
@@ -21,13 +21,20 @@ class Ui_Dialog(object):
         Dialog.setModal(True)
         self.verticalLayout = QtWidgets.QVBoxLayout(Dialog)
         self.verticalLayout.setObjectName("verticalLayout")
-        self.toolBox = QtWidgets.QToolBox(Dialog)
-        self.toolBox.setObjectName("toolBox")
+        self.tab_box = QtWidgets.QToolBox(Dialog)
+        self.tab_box.setObjectName("tab_box")
         self.page0 = QtWidgets.QWidget()
         self.page0.setGeometry(QtCore.QRect(0, 0, 420, 345))
         self.page0.setObjectName("page0")
         self.verticalLayout_2 = QtWidgets.QVBoxLayout(self.page0)
         self.verticalLayout_2.setObjectName("verticalLayout_2")
+        self.main_expr_label = QtWidgets.QLabel(self.page0)
+        self.main_expr_label.setObjectName("main_expr_label")
+        self.verticalLayout_2.addWidget(self.main_expr_label)
+        self.main_expr = QtWidgets.QLineEdit(self.page0)
+        self.main_expr.setReadOnly(True)
+        self.main_expr.setObjectName("main_expr")
+        self.verticalLayout_2.addWidget(self.main_expr)
         self.storage_label = QtWidgets.QLabel(self.page0)
         self.storage_label.setObjectName("storage_label")
         self.verticalLayout_2.addWidget(self.storage_label)
@@ -36,9 +43,9 @@ class Ui_Dialog(object):
         self.verticalLayout_2.addWidget(self.storage_list)
         icon1 = QtGui.QIcon()
         icon1.addPixmap(QtGui.QPixmap(":/icons/mechanism.png"), QtGui.QIcon.Normal, QtGui.QIcon.Off)
-        self.toolBox.addItem(self.page0, icon1, "")
+        self.tab_box.addItem(self.page0, icon1, "")
         self.page1 = QtWidgets.QWidget()
-        self.page1.setGeometry(QtCore.QRect(0, 0, 211, 135))
+        self.page1.setGeometry(QtCore.QRect(0, 0, 420, 345))
         self.page1.setObjectName("page1")
         self.horizontalLayout_2 = QtWidgets.QHBoxLayout(self.page1)
         self.horizontalLayout_2.setObjectName("horizontalLayout_2")
@@ -62,9 +69,9 @@ class Ui_Dialog(object):
         self.horizontalLayout_2.addLayout(self.verticalLayout_4)
         icon2 = QtGui.QIcon()
         icon2.addPixmap(QtGui.QPixmap(":/icons/motor.png"), QtGui.QIcon.Normal, QtGui.QIcon.Off)
-        self.toolBox.addItem(self.page1, icon2, "")
+        self.tab_box.addItem(self.page1, icon2, "")
         self.page2 = QtWidgets.QWidget()
-        self.page2.setGeometry(QtCore.QRect(0, 0, 237, 135))
+        self.page2.setGeometry(QtCore.QRect(0, 0, 420, 345))
         self.page2.setObjectName("page2")
         self.horizontalLayout_3 = QtWidgets.QHBoxLayout(self.page2)
         self.horizontalLayout_3.setObjectName("horizontalLayout_3")
@@ -88,9 +95,9 @@ class Ui_Dialog(object):
         self.horizontalLayout_3.addLayout(self.verticalLayout_5)
         icon3 = QtGui.QIcon()
         icon3.addPixmap(QtGui.QPixmap(":/icons/collections.png"), QtGui.QIcon.Normal, QtGui.QIcon.Off)
-        self.toolBox.addItem(self.page2, icon3, "")
+        self.tab_box.addItem(self.page2, icon3, "")
         self.page3 = QtWidgets.QWidget()
-        self.page3.setGeometry(QtCore.QRect(0, 0, 111, 133))
+        self.page3.setGeometry(QtCore.QRect(0, 0, 420, 345))
         self.page3.setObjectName("page3")
         self.verticalLayout_7 = QtWidgets.QVBoxLayout(self.page3)
         self.verticalLayout_7.setObjectName("verticalLayout_7")
@@ -102,8 +109,8 @@ class Ui_Dialog(object):
         self.verticalLayout_7.addWidget(self.results_list)
         icon4 = QtGui.QIcon()
         icon4.addPixmap(QtGui.QPixmap(":/icons/dimensional_synthesis.png"), QtGui.QIcon.Normal, QtGui.QIcon.Off)
-        self.toolBox.addItem(self.page3, icon4, "")
-        self.verticalLayout.addWidget(self.toolBox)
+        self.tab_box.addItem(self.page3, icon4, "")
+        self.verticalLayout.addWidget(self.tab_box)
         self.horizontalLayout = QtWidgets.QHBoxLayout()
         self.horizontalLayout.setObjectName("horizontalLayout")
         spacerItem = QtWidgets.QSpacerItem(40, 20, QtWidgets.QSizePolicy.Expanding, QtWidgets.QSizePolicy.Minimum)
@@ -120,14 +127,15 @@ class Ui_Dialog(object):
 
     def retranslateUi(self, Dialog):
         _translate = QtCore.QCoreApplication.translate
+        self.main_expr_label.setText(_translate("Dialog", "Current:"))
         self.storage_label.setText(_translate("Dialog", "Storage:"))
-        self.toolBox.setItemText(self.toolBox.indexOf(self.page0), _translate("Dialog", "Mechanism"))
+        self.tab_box.setItemText(self.tab_box.indexOf(self.page0), _translate("Dialog", "Mechanism"))
         self.variables_label.setText(_translate("Dialog", "Variables:"))
         self.records_label.setText(_translate("Dialog", "Records:"))
-        self.toolBox.setItemText(self.toolBox.indexOf(self.page1), _translate("Dialog", "Inputs"))
+        self.tab_box.setItemText(self.tab_box.indexOf(self.page1), _translate("Dialog", "Inputs"))
         self.structures_label.setText(_translate("Dialog", "Structures:"))
         self.triangular_iteration_label.setText(_translate("Dialog", "Triangular iteration:"))
-        self.toolBox.setItemText(self.toolBox.indexOf(self.page2), _translate("Dialog", "Collections"))
+        self.tab_box.setItemText(self.tab_box.indexOf(self.page2), _translate("Dialog", "Collections"))
         self.results_label.setText(_translate("Dialog", "Results:"))
-        self.toolBox.setItemText(self.toolBox.indexOf(self.page3), _translate("Dialog", "Dimensional Synthesis"))
+        self.tab_box.setItemText(self.tab_box.indexOf(self.page3), _translate("Dialog", "Dimensional Synthesis"))
 from pyslvs_ui import icons_rc