+++ docs/changelog.md
@@ -10,6 +10,7 @@
 + Add "auto remove empty links after deleted points" option. (github issue #23)
 + Move background option into project file. (github issue #26)
 + The background image can be set by dragging in.
++ Show the current mechanism in the overview dialog.  (github issue #25)
 
 **Development**: